requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argilla==1.6.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
-e git+https://github.com/argosopentech/argos-translate.git@000aecda6fc927250a1b477186aae5812fd827d0#egg=argostranslate
arrow==1.2.3
asteroid-filterbanks==0.4.0
astroid @ file:///tmp/abs_e5wkt48jiz/croots/recipe/astroid_1659023120113/work
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
backcall==0.2.0
backoff==2.2.1
backports.cached-property==1.0.2
beautifulsoup4==4.11.1
bleach==6.0.0
blobfile==2.0.1
cachetools==5.2.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.20
cmaes==0.9.1
cmake==3.26.3
colorama==0.4.6
colorlog==6.7.0
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.7
ctranslate2==2.24.0
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill @ file:///croot/dill_1667919592856/work
dnspython==1.16.0
docopt==0.6.2
duckdb==0.7.1
einops==0.3.2
entrypoints==0.4
et-xmlfile==1.1.0
eventlet==0.30.2
executing==1.2.0
faiss-gpu==1.7.1
fastapi==0.95.1
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.51.1
gunicorn==20.0.4
h11==0.14.0
hmmlearn==0.2.8
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.11.1
HyperPyYAML==1.1.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort @ file:///tmp/build/80754af9/isort_1628603791788/work
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
julius==0.2.7
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langchain==0.0.137
lazy-object-proxy @ file:///tmp/build/80754af9/lazy-object-proxy_1616529027849/work
librosa==0.9.2
lit==16.0.1
llvmlite==0.39.1
lxml==4.9.2
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe @ file:///opt/conda/conda-bld/mccabe_1644221741721/work
mistune==2.0.4
monotonic==1.6
more-itertools==9.0.0
mpmath==1.2.1
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
openai==0.27.2
-e git+https://github.com/openai/whisper.git@ea1c266709125f68a89acdc0c6ddbde6773eb0ed#egg=openai_whisper
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
optuna==3.1.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs @ file:///opt/conda/conda-bld/platformdirs_1662711380096/work
pooch==1.6.0
posthog==3.0.1
primePy==1.3
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
-e git+https://github.com/pyannote/pyannote-audio.git@0506ae38c3e03e7ce2cb0b869e986d259ffc75a4#egg=pyannote.audio
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.6
pyDeprecate==0.3.2
Pygments==2.14.0
pylint @ file:///tmp/abs_6fxmc66kyk/croots/recipe/pylint_1659110350161/work
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.8.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.4
python-magic==0.4.27
python-pptx==0.6.21
pytorch-lightning==1.6.5
pytorch-metric-learning==1.7.1
pytube==12.1.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.0
scipy==1.10.0
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
shellingham==1.5.0.post1
simplejson==3.18.1
singledispatchmethod==1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.11.0
soupsieve==2.3.2.post1
speechbrain==0.5.13
SQLAlchemy==1.4.46
stack-data==0.6.2
stanza==1.1.1
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.1.0
tiktoken==0.3.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
tomlkit @ file:///tmp/abs_56_0lnnq5x/croots/recipe/tomlkit_1658946880479/work
torch==2.0.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.2
torchaudio==0.13.1
torchmetrics==0.11.0
torchvision==0.15.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.25.1
triton==2.0.0
typer==0.7.0
typing-inspect==0.8.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
unstructured==0.6.1
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==11.0.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt @ file:///tmp/abs_c335821b-6e43-4504-9816-b1a52d3d3e1eel6uae8l/croots/recipe/wrapt_1657814400492/work
XlsxWriter==3.1.0
yarl==1.8.2
zipp==3.11.0
zstandard==0.21.0
